Join modules and simplify project structure
